feat(codebase-refactoring): [6.1] UIレンダリングとゲームロジックの分離
battle.goを3つのファイルに分割:
- battle.go (403行): Model構造体とInit/Updateメソッド
- battle_view.go (573行): UIレンダリングロジック
- battle_logic.go (408行): ゲームロジック

Requirements: 6.1, 6.2, 6.3を満たす:
- UIレンダリング責務とゲームロジック責務を別ファイルに分離
- battle.goは500行以下 (403行)
- 既存テストが全て通過
