+++ .kiro/specs/codebase-refactoring/tasks.md
@@ -107,8 +107,8 @@
 
 ### Phase 5: battle.goの分割
 
-- [ ] 6. バトル画面のファイル分割
-- [ ] 6.1 UIレンダリングとゲームロジックを分離する
+- [x] 6. バトル画面のファイル分割
+- [x] 6.1 UIレンダリングとゲームロジックを分離する
   - バトルUIのレンダリングロジックを別ファイルに抽出する
   - バトルロジック（ダメージ計算、状態遷移等）を分離する
   - メインのbattle.goが500行以下になることを確認する
